update readme, refactor
- update root level readme with instructions for installation and running

- move/rename files
