--- config.yaml
@@ -1,15 +1,0 @@
-datapath: "./data"
-resultpath: "./output"
-embeddingspath: "./output/embeddings.pkl"
-generate_visualizations: False
-low_memory: False
-nr_docs: 10
-document_diversity: 0.1
-prompt: ""
-min_topic_size: 100
-clustering_model: hdbscan
-use_llm: True
-llm_model: mistral-small
-dimensionality_reduction: umap
-save_intermediate_results: False
-save_embeddings: True
